+++ weblate/machinery/cyrtranslit.py
@@ -2,8 +2,16 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from weblate.lang.models import Language
 
 from .base import MachineTranslation, TranslationResultDict
+
+if TYPE_CHECKING:
+    from weblate.trans.models import Translation
 
 CYRTRANSLIT_TO_WEBLATE_LANGS = {
     # bulgarian
@@ -18,9 +26,9 @@ CYRTRANSLIT_TO_WEBLATE_LANGS = {
     # mongolian
     "mn@latin": [],
     "mn@cyrillic": ["mn"],
-    # russian
-    "ru@latin": [],
-    "ru@cyrillic": ["ru"],
+    # (belo)russian
+    "ru@latin": ["be_Latn"],
+    "ru@cyrillic": ["be", "ru"],
     # serbian
     "sr@latin": ["sr", "sr_Latn", "sr@ijekavian_Latn"],
     "sr@cyrillic": ["sr@ijekavian", "sr_Cyrl"],
@@ -43,6 +51,32 @@ class CyrTranslitTranslation(MachineTranslation):
     def download_languages(self):
         """List of supported languages."""
         return list(CYRTRANSLIT_TO_WEBLATE_LANGS.keys())
+
+    def get_default_source_language(self, translation: Translation) -> Language:
+        """Return default source language for the translation."""
+        mapped_code = self.map_language_code(translation.language.code)
+
+        # Invert alphabet
+        if mapped_code.endswith("@latin"):
+            mapped_code = mapped_code.replace("@latin", "@cyrillic")
+        else:
+            mapped_code = mapped_code.replace("@cyrillic", "@latin")
+
+        if mapped_code in CYRTRANSLIT_TO_WEBLATE_LANGS:
+            languages = Language.objects.filter(
+                code__in=CYRTRANSLIT_TO_WEBLATE_LANGS[mapped_code],
+                translation__component=translation.component,
+            )
+
+            # Try finding best fit based on the base code
+            for language in languages:
+                if language.code.startswith(translation.language.base_code):
+                    return language
+
+            # Return any if we failed to match
+            if languages:
+                return languages[0]
+        return super().get_default_source_language(translation)
 
     def download_translations(
         self,