+++ weblate/trans/models/component.py
@@ -28,6 +28,7 @@ from django.dispatch import receiver
 from django.utils.functional import cached_property
 from django.utils.html import format_html, format_html_join
 from django.utils.safestring import mark_safe
+from django.utils.text import format_lazy
 from django.utils.timezone import localtime
 from django.utils.translation import gettext, gettext_lazy, ngettext, pgettext
 from weblate_language_data.ambiguous import AMBIGUOUS
@@ -799,6 +800,23 @@ class Component(
             "Regular expression used to filter keys. This is only available for monolingual formats."
         ),
         blank=True,
+    )
+
+    secondary_language = models.ForeignKey(
+        Language,
+        verbose_name=gettext_lazy("Secondary language"),
+        help_text=format_lazy(
+            "{} {}",
+            gettext_lazy(
+                "Additional language to show together with the source language while translating."
+            ),
+            gettext_lazy("This setting is inherited from the project if left empty."),
+        ),
+        default=None,
+        blank=True,
+        null=True,
+        related_name="component_secondary_languages",
+        on_delete=models.deletion.CASCADE,
     )
 
     objects = ComponentQuerySet.as_manager()
@@ -3532,7 +3550,7 @@ class Component(
     @cached_property
     def all_flags(self):
         """Return parsed list of flags."""
-        return Flags(self.file_format_flags, self.check_flags)
+        return Flags(self.project.check_flags, self.file_format_flags, self.check_flags)
 
     @property
     def is_multivalue(self):