+++ weblate/trans/models/project.py
@@ -29,6 +29,7 @@ from weblate.memory.tasks import import_memory
 from weblate.trans.actions import ActionEvents
 from weblate.trans.defines import PROJECT_NAME_LENGTH
 from weblate.trans.mixins import CacheKeyMixin, LockMixin, PathMixin
+from weblate.trans.validators import validate_check_flags
 from weblate.utils.data import data_dir
 from weblate.utils.site import get_site_url
 from weblate.utils.stats import ProjectLanguage, ProjectStats, prefetch_stats
@@ -252,6 +253,27 @@ class Project(models.Model, PathMixin, CacheKeyMixin, LockMixin):
             "for example: en_GB:en,en_US:en"
         ),
         validators=[validate_language_aliases],
+    )
+    secondary_language = models.ForeignKey(
+        Language,
+        verbose_name=gettext_lazy("Secondary language"),
+        help_text=gettext_lazy(
+            "Additional language to show together with the source language while translating."
+        ),
+        default=None,
+        blank=True,
+        null=True,
+        related_name="project_secondary_languages",
+        on_delete=models.deletion.CASCADE,
+    )
+    check_flags = models.TextField(
+        verbose_name=gettext_lazy("Translation flags"),
+        default="",
+        help_text=gettext_lazy(
+            "Additional comma-separated flags to influence Weblate behavior."
+        ),
+        validators=[validate_check_flags],
+        blank=True,
     )
 
     machinery_settings = models.JSONField(default=dict, blank=True)