+++ docs/changes.rst
@@ -9,6 +9,9 @@ Weblate 5.11
 * Added :http:delete:`/api/groups/(int:id)/roles/(int:role_id)` to delete a role from a group.
 * * :ref:`addon-weblate.webhook.webhooks` are now available as a add-on.
 * :ref:`check-automattic-components-format` check to validate placeholders in Automattic components.
+* Inherited flags can now be discarded, see :ref:`custom-checks`.
+* :ref:`secondary-languages` can now be specified in :ref:`project` and :ref:`component`.
+* :ref:`mt-sources` can now be customized.
 
 .. rubric:: Improvements
 
@@ -21,7 +24,9 @@ Weblate 5.11
 * :doc:`/admin/machine` documentation extended.
 * :ref:`addon-weblate.discovery.discovery` better handles hundredths of matches.
 * Dismissing :ref:`checks` automatically updates propagated strings.
+* :ref:`project-check_flags` can now also be configured on the project level.
 * Improved rendering of :ref:`additional-flags` and :ref:`additional-explanation` changes in history.
+* :ref:`mt-cyrtranslit` now automatically transliterates from a matching translation instead of the source strings.
 
 .. rubric:: Bug fixes